requirements.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
dj-static==0.0.6
Django==1.11.6
django-cors-headers==1.1.0
django-cors-middleware==1.3.1
django-easy-select2==1.3.3
django-environ==0.4.3
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.4.7
djangorestframework-jsonapi==2.3.1
djangorestframework-jwt==1.8.0
idna==2.6
inflection==0.3.1
Markdown==2.6.6
olefile==0.44
packaging==16.8
pbr==1.10.0
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.4.2
pyparsing==2.2.0
pytz==2017.2
requests==2.18.4
six==1.10.0
static3==0.7.0
urllib3==1.22
